requirements.txt
codecov<2.1,>=2.0.5
coverage<4.4,>=4.3.4
Django<1.12,>=1.11.5
django-environ<0.5,>=0.4.4
djangorestframework<3.7,>=3.6.2
djangorestframework-jwt<1.12.0,>=1.11.0
django-suit<0.3,>=0.2.25
factory-boy<2.9,>=2.8.1
mock<2.1,>=2.0.0
Pillow<4.2,>=4.1.0
psycopg2<2.8,>==2.7.3.1
requests